Promote IrisConvexSetMaker to the namespace geometry::optimization::internal.
