+++ planning/iris/iris_np2.cc
@@ -24,6 +24,7 @@ using geometry::Role;
 using geometry::SceneGraphInspector;
 using geometry::optimization::ConvexSet;
 using geometry::optimization::internal::ClosestCollisionProgram;
+using geometry::optimization::internal::IrisConvexSetMaker;
 using geometry::optimization::internal::SamePointConstraint;
 using math::RigidTransform;
 using multibody::MultibodyPlant;
@@ -52,83 +53,6 @@ using geometry::optimization::HPolyhedron;
 using geometry::optimization::Hyperellipsoid;
 using geometry::optimization::MinkowskiSum;
 using geometry::optimization::VPolytope;
-
-// Constructs a ConvexSet for each supported Shape and adds it to the set.
-class IrisConvexSetMaker final : public ShapeReifier {
- public:
-  DRAKE_NO_COPY_NO_MOVE_NO_ASSIGN(IrisConvexSetMaker);
-
-  IrisConvexSetMaker(const QueryObject<double>& query,
-                     std::optional<FrameId> reference_frame)
-      : query_{query}, reference_frame_{reference_frame} {};
-
-  void set_reference_frame(const FrameId& reference_frame) {
-    DRAKE_DEMAND(reference_frame.is_valid());
-    *reference_frame_ = reference_frame;
-  }
-
-  void set_geometry_id(const GeometryId& geom_id) { geom_id_ = geom_id; }
-
-  using ShapeReifier::ImplementGeometry;
-
-  void ImplementGeometry(const Box&, void* data) {
-    DRAKE_DEMAND(geom_id_.is_valid());
-    auto& set = *static_cast<copyable_unique_ptr<ConvexSet>*>(data);
-    // Note: We choose HPolyhedron over VPolytope here, but the IRIS paper
-    // discusses a significant performance improvement using a "least-distance
-    // programming" instance from CVXGEN that exploited the VPolytope
-    // representation.  So we may wish to revisit this.
-    set = std::make_unique<HPolyhedron>(query_, geom_id_, reference_frame_);
-  }
-
-  void ImplementGeometry(const Capsule&, void* data) {
-    DRAKE_DEMAND(geom_id_.is_valid());
-    auto& set = *static_cast<copyable_unique_ptr<ConvexSet>*>(data);
-    set = std::make_unique<MinkowskiSum>(query_, geom_id_, reference_frame_);
-  }
-
-  void ImplementGeometry(const Cylinder&, void* data) {
-    DRAKE_DEMAND(geom_id_.is_valid());
-    auto& set = *static_cast<copyable_unique_ptr<ConvexSet>*>(data);
-    set =
-        std::make_unique<CartesianProduct>(query_, geom_id_, reference_frame_);
-  }
-
-  void ImplementGeometry(const Ellipsoid&, void* data) {
-    DRAKE_DEMAND(geom_id_.is_valid());
-    auto& set = *static_cast<copyable_unique_ptr<ConvexSet>*>(data);
-    set = std::make_unique<Hyperellipsoid>(query_, geom_id_, reference_frame_);
-  }
-
-  void ImplementGeometry(const HalfSpace&, void* data) {
-    DRAKE_DEMAND(geom_id_.is_valid());
-    auto& set = *static_cast<copyable_unique_ptr<ConvexSet>*>(data);
-    set = std::make_unique<HPolyhedron>(query_, geom_id_, reference_frame_);
-  }
-
-  void ImplementGeometry(const Sphere&, void* data) {
-    DRAKE_DEMAND(geom_id_.is_valid());
-    auto& set = *static_cast<copyable_unique_ptr<ConvexSet>*>(data);
-    set = std::make_unique<Hyperellipsoid>(query_, geom_id_, reference_frame_);
-  }
-
-  void ImplementGeometry(const Convex&, void* data) {
-    DRAKE_DEMAND(geom_id_.is_valid());
-    auto& set = *static_cast<copyable_unique_ptr<ConvexSet>*>(data);
-    set = std::make_unique<VPolytope>(query_, geom_id_, reference_frame_);
-  }
-
-  void ImplementGeometry(const Mesh&, void* data) {
-    DRAKE_DEMAND(geom_id_.is_valid());
-    auto& set = *static_cast<copyable_unique_ptr<ConvexSet>*>(data);
-    set = std::make_unique<VPolytope>(query_, geom_id_, reference_frame_);
-  }
-
- private:
-  const QueryObject<double>& query_{};
-  std::optional<FrameId> reference_frame_{};
-  GeometryId geom_id_{};
-};
 
 struct GeometryPairWithDistance {
   GeometryId geomA;
